user_management/requirements.txt
fastapi
uvicorn
tortoise-orm
redis
cryptography
pydantic_settings
pika
aiomysql
aio_pika
rich
passlib